New: New dir-table-filter component - Updated styles for directory route
