+++ packages/directory/addon/templates/directory.hbs
@@ -1,11 +1,28 @@
 {{!-- Copyright 2018, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
 <main class='directory'>
+  
   <div class='directory__new-button'>
-    {{dir-new-button class='directory__new-button'}}
+    {{dir-new-button classNames='directory__new-button'}}
   </div>
-  {{dir-search-bar class='directory__search' query=q searchFor=(action 'searchFor')}}
-  {{dir-sidebar class='directory__sidebar'}}
+  
+  <section class='directory__controls'>
+    {{dir-search-bar 
+      classNames='directory__search' 
+      query=q 
+      searchFor=(action 'searchFor')
+    }}
+
+    {{dir-table-filter
+      classNames='directory__table-filter'
+      selectedType=type
+      updateQueryParams=(route-action 'updateQueryParams')
+    }}
+  </section>
+  
+  {{dir-sidebar classNames='directory__sidebar'}}
+
   <section class='directory__table'>
     {{outlet}}
   </section>
+
 </main>
