+++ packages/directory/addon/templates/directory/my-directory.hbs
@@ -2,6 +2,4 @@
 {{dir-table
   items=searchResults
   searchQuery=directory.q
-  typeFilter=directory.type
-  updateQueryParams=(route-action 'updateQueryParams')
 }}
