+++ packages/directory/app/components/dir-table-filter.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-directory/components/dir-table-filter';
